IT-9
I want to be able to create an account that contains.
         a) Name
         b) Contact information (Email and Phone)
          c) AutoGenerated Membership number

The user name should always be the users email.

The password created by the user need to follow:

a) At least 8 characters

b) A mixture of both uppercase and lowercase letters.
c) A mixture of letters and numbers.
d) Inclusion of at least one special character, e.g., ! @ # ? ]

http://localhost:3000/register